refactor(site): move the WebGL renderer probe into gates/
gates/ was meant to be the self-contained capability-policy layer, but
its aurora policy imported the probe from the landing root - the only
one of the four WebGL capability files living outside gates/. The probe
and the software-renderer classifier are a pair (detect + classify), so
they now sit together and aurora-gate's dependency is intra-layer.
Import sites updated; no behavior change.
